requirements.txt
openai
moviepy
google-cloud-texttospeech
google-cloud-speech
google-api-python-client
google-auth-oauthlib
requests